Remove 'eyebrowse' local dep. Shouldn't clutter this lib with experiments
I made a messy mistake here, and have to adopt a more disciplined
approach to this library. Basically, I'll keep 'main' safe to use, and
experiment on other branches. This may be obvious to devs, but hey,
I'm a slow learner sometimes.

Anyway, this should clean things up a bit.